+++ src/signals/incident-management/containers/Dashboard/components/AreaChart/types.ts
@@ -1,0 +1,9 @@
+export enum Direction {
+  UP = 'up',
+  DOWN = 'down',
+}
+
+export interface ComparisonRateType {
+  direction: Direction
+  percentage: number
+}
