+++ src/signals/incident-management/containers/Dashboard/components/ModuleTitle/index.ts
@@ -1,0 +1,1 @@
+export { ModuleTitle } from './ModuleTitle'
